feat: 회원가입 api 생성

+++ be/src/main/java/todo/todo/dto/object/FollowList.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class FollowList {
-    private Long followerId;   // 팔로우 하는 사용자 ID
-    private Long followingId;  // 팔로우 당하는 사용자 ID
+    private int followerId;   // 팔로우 하는 사용자 ID
+    private int followingId;  // 팔로우 당하는 사용자 ID
 }
